Update and rename simple.py to simple_slow.py
I'd like to use simple.py to read SBP via a datalogger I want to build. It will be Bluetooth receiver and a small processor board with a screen and keyboard, all in a waterproof case. It will receive data from an HC-06 Bluetooth (BT) already transmitting inside the Rover waterproof case.